test-requirements.txt
--extra-index-url=https://wheelhouse.shopinvader.com/simple
odoo_test_helper
odoo10-addon-base_search_custom_field_filter @ git+https://github.com/OCA/server-ux@refs/pull/170/head#subdirectory=setup/base_search_custom_field_filter
odoo10-addon-mass-editing @ https://github.com/OCA/server-tools/refs/pull/1856/head#subdirectory=setup/mass_editing
